refactor: separate use cases
